split edge module in submodules
