+++ libafl_qemu/src/modules/edges/classic.rs
@@ -1,0 +1,113 @@
+use libafl::{inputs::UsesInput, HasMetadata};
+
+use crate::{
+    modules::{
+        helpers::{
+            gen_hashed_block_ids, trace_block_transition_hitcount, trace_block_transition_single,
+        },
+        AddressFilter, EdgeCoverageModule, EdgeCoverageModuleBuilder, EdgeCoverageVariant,
+        EmulatorModuleTuple, PageFilter, StdAddressFilter, StdPageFilter,
+    },
+    EmulatorModules, Hook,
+};
+
+#[derive(Debug)]
+pub struct EdgeCoverageClassicVariant;
+
+pub type StdEdgeCoverageClassicModule =
+    EdgeCoverageModule<StdAddressFilter, StdPageFilter, EdgeCoverageClassicVariant>;
+pub type StdEdgeCoverageClassicModuleBuilder =
+    EdgeCoverageModuleBuilder<StdAddressFilter, StdPageFilter, EdgeCoverageClassicVariant, false>;
+
+impl<AF, PF> EdgeCoverageVariant<AF, PF> for EdgeCoverageClassicVariant {
+    const DO_SIDE_EFFECTS: bool = false;
+
+    fn jit_hitcount<ET, S>(&mut self, emulator_modules: &mut EmulatorModules<ET, S>)
+    where
+        AF: AddressFilter,
+        ET: EmulatorModuleTuple<S>,
+        PF: PageFilter,
+        S: Unpin + UsesInput + HasMetadata,
+    {
+        let hook_id = emulator_modules.blocks(
+            Hook::Function(gen_hashed_block_ids::<AF, ET, PF, S, Self>),
+            Hook::Empty,
+            Hook::Empty,
+        );
+
+        unsafe {
+            libafl_qemu_sys::libafl_qemu_block_hook_set_jit(
+                hook_id.0,
+                Some(libafl_qemu_sys::libafl_jit_trace_block_hitcount),
+            );
+        }
+    }
+
+    fn jit_no_hitcount<ET, S>(&mut self, emulator_modules: &mut EmulatorModules<ET, S>)
+    where
+        AF: AddressFilter,
+        ET: EmulatorModuleTuple<S>,
+        PF: PageFilter,
+        S: Unpin + UsesInput + HasMetadata,
+    {
+        let hook_id = emulator_modules.blocks(
+            Hook::Function(gen_hashed_block_ids::<AF, ET, PF, S, Self>),
+            Hook::Empty,
+            Hook::Empty,
+        );
+
+        unsafe {
+            libafl_qemu_sys::libafl_qemu_block_hook_set_jit(
+                hook_id.0,
+                Some(libafl_qemu_sys::libafl_jit_trace_block_single),
+            );
+        }
+    }
+
+    fn fn_hitcount<ET, S>(&mut self, emulator_modules: &mut EmulatorModules<ET, S>)
+    where
+        AF: AddressFilter,
+        ET: EmulatorModuleTuple<S>,
+        PF: PageFilter,
+        S: Unpin + UsesInput + HasMetadata,
+    {
+        emulator_modules.blocks(
+            Hook::Function(gen_hashed_block_ids::<AF, ET, PF, S, Self>),
+            Hook::Empty,
+            Hook::Raw(trace_block_transition_hitcount),
+        );
+    }
+
+    fn fn_no_hitcount<ET, S>(&mut self, emulator_modules: &mut EmulatorModules<ET, S>)
+    where
+        AF: AddressFilter,
+        ET: EmulatorModuleTuple<S>,
+        PF: PageFilter,
+        S: Unpin + UsesInput + HasMetadata,
+    {
+        emulator_modules.blocks(
+            Hook::Function(gen_hashed_block_ids::<AF, ET, PF, S, Self>),
+            Hook::Empty,
+            Hook::Raw(trace_block_transition_single),
+        );
+    }
+}
+
+impl Default for StdEdgeCoverageClassicModuleBuilder {
+    fn default() -> Self {
+        Self {
+            variant: EdgeCoverageClassicVariant,
+            address_filter: StdAddressFilter::default(),
+            page_filter: StdPageFilter::default(),
+            use_hitcounts: true,
+            use_jit: true,
+        }
+    }
+}
+
+impl StdEdgeCoverageClassicModule {
+    #[must_use]
+    pub fn builder() -> StdEdgeCoverageClassicModuleBuilder {
+        EdgeCoverageModuleBuilder::default()
+    }
+}
